:recycle: (useDistance) move to appropriate location
