+++ src/features/cheatcodes/pages/Navigation.tsx
@@ -10,7 +10,7 @@ import { useSignIn } from 'features/auth/api'
 import { DEEPLINK_DOMAIN } from 'features/deeplinks'
 import { openExternalUrl } from 'features/navigation/helpers'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
-import { useDistance } from 'features/offer/components/useDistance'
+import { useDistance } from 'libs/geolocation/hooks/useDistance'
 import { AsyncError } from 'libs/errorMonitoring'
 import { QueryKeys } from 'libs/queryKeys'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
