+++ src/features/favorites/atoms/Favorite.tsx
@@ -11,7 +11,7 @@ import { useRemoveFavorite } from 'features/favorites/pages/useFavorites'
 import { mergeOfferData } from 'features/home/atoms/OfferTile'
 import { Credit } from 'features/home/services/useAvailableCredit'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
-import { useDistance } from 'features/offer/components/useDistance'
+import { useDistance } from 'libs/geolocation/hooks/useDistance'
 import { OfferImage } from 'features/search/atoms/OfferImage'
 import { CATEGORY_CRITERIA } from 'features/search/enums'
 import { formatToFrenchDate, getFavoriteDisplayPrice, parseCategory } from 'libs/parsers'
