+++ src/features/search/atoms/Hit.tsx
@@ -6,9 +6,9 @@ import styled from 'styled-components/native'
 
 import { mergeOfferData } from 'features/home/atoms/OfferTile'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
-import { useDistance } from 'features/offer/components/useDistance'
 import { CATEGORY_CRITERIA } from 'features/search/enums'
 import { analytics } from 'libs/analytics'
+import { useDistance } from 'libs/geolocation/hooks/useDistance'
 import { formatDates, getDisplayPrice, parseCategory } from 'libs/parsers'
 import { SearchHit } from 'libs/search'
 import { ColorsEnum, getSpacing, Spacer, Typo } from 'ui/theme'
